fix(market): address code review issues for city comparison
- Use Literal types for area_type and trend in schemas (#1)
- Guard bestYield against empty yield arrays (#5)
- Fix keys.sort() mutating input array in query key factory (#6)
- Add click-outside handler to close AreaSelector dropdown (#4)
- Extract CityComparison into folder with sub-components (#7)
