+++ frontend/src/query/queryKeys.ts
@@ -91,7 +91,7 @@ export const queryKeys = {
       ] as const,
     areas: () => [...queryKeys.market.all, "areas"] as const,
     cityComparison: (keys: string[]) =>
-      [...queryKeys.market.all, "cityComparison", ...keys.sort()] as const,
+      [...queryKeys.market.all, "cityComparison", ...[...keys].sort()] as const,
   },
 
   // Financing queries
